Driver Safety Prediction using ML Algorithms
In this Machine Learning Project I used various techniques such as data normalization, feature scaling, and dimensionality reduction to prepare the data for training the machine learning model. and the selected algorithm was trained on the preprocessed data using various hyperparameters to optimize its performance. The trained model was then tested on a separate dataset to evaluate its accuracy. My machine learning model achieved an accuracy of 85% in predicting driver safety. The model was able to identify risky driving behavior such as sudden braking, speeding, and lane changing